Move NativeBuildQuestItem to GameEnvironment because it depends on specific types
